// src/types/SanityResults.ts
export type ProductVersion = {
  _id: string;
  _type: "productVersion";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  versionName: string;
  description: string[];
  imgUrl: string;
  relatedProduct?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "products";
  };
};
// ...
export type Products = {
  _id: string;
  _type: "products";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title: string;
  description: string[];
  imgUrl: string;
  thumbUrl: string;
  versions: ProductVersion[];
  tags?: Array<string>;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;